fix: Rename rule Renamed rule single-space-colon to no-leading-space-in-subject
Change-type: major
