+++ doc/README.hbs
@@ -51,9 +51,9 @@ Signed-off-by: Fake Name <fakename@resin.io>
 # Rules
 
 {{#identifiers}}
-### {{name}}
+## {{name}}
 *Default: {{default}}*
 
-{{description}}
+{{{description}}}
 
 {{/identifiers}}
