+++ lib/errors.js
@@ -110,12 +110,34 @@ module.exports.CAPITALISE_SUBJECT = Error('The commit subject must start with a 
 module.exports.NO_PERIOD = Error('The commit title should not end with a period')
 
 /**
-* @name single-space-colon
-* @description The commit title must include exactly one space after the colon
+* @name no-leading-space-in-subject
+* @description
+* The commit subject must not start with a leading space.
+*
+* Note that if a prefix is supplied a space is expected between the colon and the title
+*
+* Accepts:
+*
+* ```
+* prefix: subject
+* ```
+* ```
+* subject
+* ```
+*
+* Rejects:
+*
+* ```
+* prefix:  subject
+* ```
+* ```
+*  subject
+* ```
+*
 * @const
 * @public
 */
-module.exports.SINGLE_SPACE_COLON = Error('The commit title must include exactly one space after the colon')
+module.exports.NO_LEADING_SPACE_IN_SUBJECT = Error('The commit subject must not start with a leading space')
 
 /**
 * @name imperative-mood
